Check for trivial case in findInvisibleMember
Check if 'what' is ancestor of 'from' in terms of containing declaraions
in 'Visibilities.findInvisibleMember', and if true, do not launch whole
visibility checking algorithm.

This particulary helps for cases like this:
```
class Irrelevant {
    protected open class A {
        object B : A()
    }
}
```
The problem here is that we encounter recursion during resolution of
'B':
1. For resolving B, we have to resolve it's supertypes, i.e. A
2. We resolve A successfully, but then try to check if A is visible from
B
3. We call into 'findInvisibleMember', and for 'PROTECTED' visibility it
involves access to supertypes of B, so we get a bogus recursion here.