Uast: `KotlinIDERenderLogTest` for running uast tests with IDE-setup
under IDE-setup Ultra Light Classes are implied.
In case of result difference the `-ide`-suffixed testdata is used.
Actually the difference in results is more likely to be a bug
but we currently we are fixating at least some behaviour to protect from
 unexpected regressions.

`object Local` were removed because there is no local objects in actual
 Kotlin syntax