Make outputs' backup more granular & lazy
The fix for KT-12295 added a backup of KotlinCompile task's outputs.
All outputs (IC caches & the compiler's outputs) were read to
the memory & restored in case of compilation error.
It was done to workaround Gradle forcing (providing 'unknown' inputs
changes) a non-incremental execution
if a task changed its outputs & reported an error.

A complete backup of all outputs is often inefficient.
For example a task might only have changes in its dependencies.
When the task analyzes changes in dependencies, it might determine
that its source files are not affected, so the backup was a complete waste.

This commit introduces the following changes in order to address that issue:
1. The compiler's outputs are backed up lazily.
Only dirty output files are backed up before each iteration.
2. IC caches are opened twice:
2.1. In read-only mode to determine affected files,
when analyzing dependencies, Java files, etc.
Read-only mode is implemented by wrapping LazyStorage.
LazyStorage is a wrapper for IJ PersistentHashTable
(IC does not access PHT API directly).
2.2. In the main IC loop. At this point all caches for the task
are backed up in the same way as before. The backup could be more granular,
however that would not be very useful, because usually the IC
uses most of the caches after an iteration (updating lookups, protos, contants values, etc.).

Further improvements can be made:
* Instead of backing up the caches, we could store all changes in memory,
writing only after successful incremental compilation.
* Instead of backing up the compiler's outputs, we could implement a simple
VFS for reading/writing outputs. This way we could avoid touching outputs at all.
A deletion would remove an entry from VFS instead of actually deleting a file.
A write would write file's contents to a buffer (or directly to the disk
in case of non-incremental build).