Create kotlinComponents for all Android variants, even non-published
The `kotlinComponents` are used not just in publishing but also in
source sets visibilty inference in HMPP. If a compilation is omitted
in the `kotlinComponents` and the project dependency gets resolved to
it, then source set visibility won't work properly as it won't get the
compilation's variant from the project and will thus consider the
dependency unresolved, not choosing any source sets.

To fix this, create the `kotlinComponents` from all Android variants,
not just the published ones, but set `publishable = false` in the
variants that the plugin should not create the publications for. The
plugin checks this flag already in `createMavenPublications`.