Fix stdlib-by-default with non-compiled source sets (KT-40559)
In `com.android.library` projects, there's no androidTest variants, but
AGP still creates the `androidTest*` source sets, which are then unused
and not included into any compilation.

For source sets that are not included into any compilation, the stdlib
module added by default was kotlin-stdlib-common, and it was added to
the `api` scope. But AGP checks the `androidTest*Api` configurations and
if it detects any dependencies in them, it reports deprecation warnings.

However, we have plans to prohibit unused source sets at all as they
have no reasonable use cases. So the fix is not to add the stdlib by
default to source sets that participate in no compilations.

Issue #KT-40559 Fixed