KT-41295: Friend paths and compiler options compatible with conf caching
This change fixes serialization of friend paths and it uses
a file collection to record all friend paths. Also, when
computing the tested classpath for android projects, we now avoid
resolving the file collection until necessary.

In addition to this, this starts removal of cross-task references.
More specifically, KaptGenerateStubsTask is using its own
compilerArgumentsContributor and it does not rely on the
KotlinCompile one.

Fixes: KT-41295
Test: ConfigurationCacheForAndroidIT