Decrease priority of repeatedly added ivy repository
Previously, after repository addition latter additions
of the same repository changed nothing. Now, repeatedly added
repository moves to the end of repositories list and its
resolution priority decreases.

It may be used to change the existing repository priority. I.e.
if you added A, B and C repository they will be checked
for dependencies in order A -> B -> C. If you then add B and
A repo (in this exact order), the order changes to C -> B -> A.