Built-ins: provide module dependency on sdk's built-ins if absent
Some IR components expect that if descriptor is contained in a module,
that module will be in dependencies of the current scope module.
JvmBultins from dependencies violate this requirement if <builtins-module>
(built-ins module of SDK) is not provided manually for all the plain
modules with dependency-based JvmBultins. They don't depend on
<built-ins module> (their built-ins module is a standard library
module).

KT-33233